collect/requirements.txt
PyYAML
mongoengine
requests
